Fixes RemoveAllTokensAction’s target
`RemoveAllTokensAction` (used by Flying Broomstick) creates
`RemoveTokenAction`s correctly based on the selected card’s
token counts, but then applied the actions to `context.source`,
which was the triggering card, instead of the targeted card.

For example, in the case of Flying Broomstick, `context.source`
was the creature that had the Flying Broomstick upgrade and
reaped, not the creature that was targeted by the effect.

This changes the resolving to use the targeted card as its target.

Updates the Flying Broomstick tests to check stun (which is
a counter that Flying Broomstick should remove) to validate this
case, since the æmber removal and damage healing are handled by other
game actions that weren’t buggy.

Also changes to set damage directly before triggering Flying Broomstick,
for clarity.