Adds while conditional to lasting effects, fixes token effects
Fixes #4688

This addresses bugs in how Vial of Mutation, Revo Hooligans, and
Eclectic Ambrosius (Evil Twin) implemented their effects to add
the “mutant” trait, switch to house Mars, or blank the text box,
respectively.

The previous implementations for the first two of those cards did not
consistently remove their effects when the tokens were removed, since
they listened to `onAddToken` instead of `onRemoveToken`. Additionally,
they did not correctly implement that losing the token but then
regaining it should restore the effect, based on the FAQ entry for
Sneklifter’s similarly lasting effect.

Eclectic Ambrosius (Evil Twin) incorrectly applied the text box–blanking
effect to any creature that got an ignorance counter, not just ones
targeted by the “After Reap:” effect. This would have incorrectly
blanked Envoy of Ekwirrĕ’s text box if that card swapped the ignorance
counter on to itself.

To manage the “this effect is not applied but is waiting in case it should be
re-applied” behavior, this adds a `while` predicate to `CardLastingEffectAction`
and `Effect`. If a `while` function is provided, otherwise-valid targets for whom
`while` returns `false` have the effect un-applied. They will get it re-applied
if later `while` returns `true` for them.

Though this behavior is simliar to the effects of returning `false` from `isValidTarget`,
it is kept as a separate operation because the targets are still valid for the overall
effect, they just don’t meet the criteria to have the effect applied.