Fixes CopyCard’s handling of copying underlying card
Fixes #3529

This handles the case where cards flipped face-up by Temporal
Purge were maintaining abilities from their token side. Flipping
face-up is done by doing a CopyCard to overwrite the CopyCard
that gave the card the token abilities. The bug in the previous
implementation was that the source for the “original card” was
`Card.abilities`, which was including both the original card’s
abilities as well as those added by the previous CopyCard.

This adjusts that logic to explicitly filter out any abilities that
came from CopyCard effects.

Adds test to Temporal Purge, as well as to Mirror Shell to capture
that the behavior is correct through 2 layers of copying.

Also renames `cascadeEffects` to `copyOriginalCard` for clarity,
inverting its value (so to copy the original you pass `true` rather
than `false`).